backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets>=13.1,<14.0
python-dotenv==1.0.0
hume
